Experimental change to fix handling of interim responses
when the illumos server responds STATUS_PENDING to some command that's
part of a compount, that command AND all subsequent commands in that
compound are given an interim response.  Before this change, the
client bails out on the compounded interim responses.

This fix is probably not quite right, but mostly works.